Response was not captured for the gzip content and extraction of data from the gzip file was not possible - Fixed.
In Extraction Failed situation, extracted variables could not be used in condition statement and JS file - Fixed. Now when extraction fails, variables is updated with "NOTFOUND" static text and this text can be used for validating in conditional statement or in JS files.

Thread aborted error on if condition due to improper validation is fixed. Thread now will abort only on abnormal situation like instead of giving variable with $$ giving with direct name will still fail.