Below Fixes Done to the build AppedoLT Version 3.3.042 & LT generator Version 1.2.044
Container Response time is calculated as difference from Max(endtime) -
Min(Starttime) to take care parallel connection effect and it is done at page level then moved to container level to avoid think time.

Previously it was sum(diff) of all request within container.

When content type chunked and not a gzip, response was not captured. This was causing extraction failure error during extraction. This is resolved.

Thread pool concept introduced where ever parallel threads are used to ensure thread overhead is avoided while doing multiple iterations or run in duration. This will considerably reduce the overhead of CPU and thread creation time.

User creation is still with normal thread creation mode as it is very long thread and user count might vary for each run and it is one time activity for entire run.